Refactor existing secondary buttons to new style
