+++ src/features/ConfirmAttendance/ConfirmAttendance.tsx
@@ -7,7 +7,8 @@ import { boolean, object } from 'yup';
 
 import { Hacker } from '../../api';
 import { FrontendRoute } from '../../config';
-import { Button, H1, H2, MaxWidthBox, Paragraph } from '../../shared/Elements';
+import { H1, H2, MaxWidthBox, Paragraph } from '../../shared/Elements';
+import Button, { ButtonProminence } from '../../shared/Elements/Button';
 import { Form } from '../../shared/Form';
 import * as FormikElements from '../../shared/Form/FormikElements';
 import ValidationErrorGenerator from '../../shared/Form/validationErrorGenerator';
@@ -97,7 +98,7 @@ class ConfirmAttendanceContainer extends React.Component<
           <Box>
             <Button
               onClick={this.onClickFactory(0, fp.submitForm)}
-              secondary={true}
+              prominence={ButtonProminence.Secondary}
               isLoading={fp.isSubmitting}
               disabled={fp.isSubmitting}
             >
