+++ src/features/Nav/LogoutButton.tsx
@@ -3,9 +3,7 @@ import * as React from 'react';
 import { RouteComponentProps, withRouter } from 'react-router-dom';
 import { APIResponse, Auth } from '../../api';
 import { FrontendRoute, IValidationError } from '../../config';
-import Button, {
-  ButtonProminence,
-} from '../../shared/Elements/ButtonRefactored';
+import Button, { ButtonProminence } from '../../shared/Elements/Button';
 import ValidationErrorGenerator from '../../shared/Form/validationErrorGenerator';
 
 const LogoutBtn: React.StatelessComponent<RouteComponentProps> = (
